* master: (183 commits)
  Adding tests for writing null blocks with and without compression
  Replace LookupWriterRec's linear search on RecList with an unordered_map. For 250k variables, time goes from 21sec to ~1sec in WSL. The order of entries in RecList was not necessary for the serializer to work correctly.
  Merge pull request #3805 from pnorbert/fix-bpls-string-scalar
  Merge pull request #3804 from pnorbert/fix-aws-version
  Merge pull request #3759 from pnorbert/bp5dbg-metadata
  new attempt to commit query support of local array. (#3868)
  MPI::MPI_Fortran should be INTERFACE not PUBLIC
  Fix hip example compilation error (#3865)
  Server Improvements (#3862)
  ascent,ci: remove unshallow flag
  Remove Slack as a contact mechanism (#3866)
  bug fix:  syntax error in json  output (#3857)
  Update the bpWriterReadHip example's cmake to run on crusher
  Examples: Use BPFile instead of BP3/4/5 for future-proof
  inlineMWE example: Close files at the end
  Examples: Add BeginStep/EndStep wherever it was missing
  BP5Serializer: handle local variables that use operators (#3859)
  Blosc2 USE ON: Fix Module Fallback (#3774)
  SST,MPI,DP: soft handle peer error
  SST,MPI,DP: improve uniq identifier
  Fix destdir install test (#3850)
  cmake: include ctest before detectoptions
  ci: enable tau check
  Add/Improve the ReadMe.md files in examples directory
  Disable BUILD_TESTING and ADIOS2_BUILD_EXAMPLES by default
  Remove testing based on ADIOS2-examples
  Fix formatting issue in DetectOptions.cmake
  Add examples from ADIOS2-Examples
  Improve existing examples
  MPI_DP: do not call MPI_Init (#3847)
  cmake: update minimum cmake to 3.12 (#3849)
  MPI: add timeout for conf test for MPI_DP (#3848)
  Tweak Remote class and test multi-threaded file remote access (#3834)
  Add prototype testing of remote functionality (#3830)
  Try always using the MPI version
  Try always using the MPI version
  Import tests from bp to staging common, implement memory selection in SST
  ci: fix codeql ignore path (#3772)
  install: export adios2 device variables (#3819)
  added support to query BP5 files (#3809)
  ffs 2023-09-19 (67e411c0)
  Fix abs/rel step in BP5 DoCount
  fix dummy Win build
  Pass Array Order of reader to remote server for proper Get() operation
  Fix the ADIOS_USE_{_} variable names to use ADIOS2
  Remove BP5 BetweenStepPairs variable that hides Engine.h counterpart
  Delete experimental examples
  yaml-cpp: support 0.8.0 version
  Mod not to overload to make some compilers happy
  Mod to keep MinBlocksInfo private to BP5Writer, not reuse Engine.h API.
  gha,ci: update checkout to v4
  Limit testing changes to BP5
  tests: fix windows 64bit type issues
  Add location of hdf5.dll to PATH for windows tests
  Fix more compiler warnings from ompi build
  clang format
  Fix size_t -> int conversion warnings
  Allow building with Visual Studio and HDF5 shared libs
  Fixup local var reading by block with test, master branch
  use documented initial cache variable to find hdf5 on windows
  ci: Add HDF5 to a windows build
  Remove unused SelectionType values
  Update ADIOS2 HDF5 VOL with basic set of capability flags
  defines
  reorder
  Remove file before writing, catch exceptions
  format
  Reader-side Profiling
  removed  a comment
  ci: disable MGARD static build
  cmake: fix ffs dependency
  removed comments and cleaned up more code
  bp5: remove ADIOS2_USE_BP5 option
  operators: fix module library
  ci: Create static minimal build
  cmake: correct prefer_shared_blosc behavior
  removing warning from auto
  clang-format
  match type of timestep for h5 engine to size_t (same as adios VariableBase class) when storing to hdf5, check the limit of size_t and match to the right type (uint, ulong, default ulonglong)
  BP5Deserialize: modify changes to keep abi compt
  Rename test. fix windows compile errors.
  add codeql workflow
  Fix ChunkV maintaining CurOffset when downsizing current chunk in Allocate() and creating the next chunk.
  ci: fix docker images names
  removed kwsys gitattributes
  KWSys 2023-05-25 (c9f0da47)
  fixup! ci: add mgard dependency to spack builds
  ci: remove unused script part
  cmake: remove redundant pugixml dependency
  ci: add mgard dependency to spack builds
  cmake: Remove enet config warning
  cmake: resolve cmake python deprecation
  enet 2023-08-15 (7587eb15)
  EVPath 2023-08-15 (657c7fa4)
  dill 2023-08-15 (235dadf0)
  atl 2023-08-15 (7286dd11)
  remove data.py since BP5 data file has no format
  format
  bp5dbg parse records and check sizes during it for mmd.0 and md.0. No decoding FFS records.
  cmake: correct info.h installation path
  ...

+++ source/adios2/engine/bp5/BP5Writer_EveryoneWrites_Async.cpp
@@ -30,8 +30,8 @@ namespace engine
 
 using namespace adios2::format;
 
-BP5Writer::ComputationStatus
-BP5Writer::IsInComputationBlock(AsyncWriteInfo *info, size_t &compBlockIdx)
+BP5Writer::ComputationStatus BP5Writer::IsInComputationBlock(AsyncWriteInfo *info,
+                                                             size_t &compBlockIdx)
 {
     ComputationStatus compStatus = ComputationStatus::NotInComp_ExpectMore;
     size_t nExpectedBlocks = info->expectedComputationBlocks.size();
@@ -54,13 +54,11 @@ BP5Writer::IsInComputationBlock(AsyncWriteInfo *info, size_t &compBlockIdx)
         if (inComp)
         {
             while (compBlockIdx < nExpectedBlocks &&
-                   info->expectedComputationBlocks[compBlockIdx].blockID <
-                       compBlockID)
+                   info->expectedComputationBlocks[compBlockIdx].blockID < compBlockID)
             {
                 ++compBlockIdx;
             }
-            if (info->expectedComputationBlocks[compBlockIdx].blockID >
-                compBlockID)
+            if (info->expectedComputationBlocks[compBlockIdx].blockID > compBlockID)
             {
                 // the current computation block is a short one that was not
                 // recorded
@@ -75,10 +73,8 @@ BP5Writer::IsInComputationBlock(AsyncWriteInfo *info, size_t &compBlockIdx)
     return compStatus;
 }
 
-void BP5Writer::AsyncWriteOwnData(AsyncWriteInfo *info,
-                                  std::vector<core::iovec> &DataVec,
-                                  const size_t totalsize,
-                                  const bool seekOnFirstWrite)
+void BP5Writer::AsyncWriteOwnData(AsyncWriteInfo *info, std::vector<core::iovec> &DataVec,
+                                  const size_t totalsize, const bool seekOnFirstWrite)
 {
     /* local variables to track variables modified by main thread */
     size_t compBlockIdx = 0; /* position in vector to get length */
@@ -88,7 +84,7 @@ void BP5Writer::AsyncWriteOwnData(AsyncWriteInfo *info,
     size_t wrote = 0;
     size_t block = 0;
     size_t temp_offset = 0;
-    size_t max_size = std::max(1024 * 1024UL, totalsize / 100UL);
+    size_t max_size = std::max((size_t)1024 * 1024UL, totalsize / 100UL);
 
     bool firstWrite = seekOnFirstWrite;
     while (block < nBlocks)
@@ -102,8 +98,7 @@ void BP5Writer::AsyncWriteOwnData(AsyncWriteInfo *info,
 
         if (!doRush)
         {
-            ComputationStatus compStatus =
-                IsInComputationBlock(info, compBlockIdx);
+            ComputationStatus compStatus = IsInComputationBlock(info, compBlockIdx);
 
             /* Scheduling decisions:
                Cases:
@@ -136,10 +131,8 @@ void BP5Writer::AsyncWriteOwnData(AsyncWriteInfo *info,
 
         if (doRush)
         {
-            auto vec = std::vector<adios2::core::iovec>(DataVec.begin() + block,
-                                                        DataVec.end());
-            vec[0].iov_base =
-                (const char *)DataVec[block].iov_base + temp_offset;
+            auto vec = std::vector<adios2::core::iovec>(DataVec.begin() + block, DataVec.end());
+            vec[0].iov_base = (const char *)DataVec[block].iov_base + temp_offset;
             vec[0].iov_len = DataVec[block].iov_len - temp_offset;
             size_t pos = MaxSizeT; // <==> no seek inside WriteFileAt
             if (firstWrite)
@@ -172,15 +165,13 @@ void BP5Writer::AsyncWriteOwnData(AsyncWriteInfo *info,
 
         if (firstWrite)
         {
-            info->tm->WriteFileAt((const char *)DataVec[block].iov_base +
-                                      temp_offset,
-                                  n, info->startPos);
+            info->tm->WriteFileAt((const char *)DataVec[block].iov_base + temp_offset, n,
+                                  info->startPos);
             firstWrite = false;
         }
         else
         {
-            info->tm->WriteFiles(
-                (const char *)DataVec[block].iov_base + temp_offset, n);
+            info->tm->WriteFiles((const char *)DataVec[block].iov_base + temp_offset, n);
         }
 
         /* Have we processed the entire block or staying with it? */
@@ -224,34 +215,29 @@ int BP5Writer::AsyncWriteThread_EveryoneWrites(AsyncWriteInfo *info)
     return 1;
 };
 
-void BP5Writer::WriteData_EveryoneWrites_Async(format::BufferV *Data,
-                                               bool SerializedWriters)
+void BP5Writer::WriteData_EveryoneWrites_Async(format::BufferV *Data, bool SerializedWriters)
 {
 
-    const aggregator::MPIChain *a =
-        dynamic_cast<aggregator::MPIChain *>(m_Aggregator);
+    const aggregator::MPIChain *a = dynamic_cast<aggregator::MPIChain *>(m_Aggregator);
 
     // new step writing starts at offset m_DataPos on aggregator
     // others will wait for the position to arrive from the rank below
 
     if (a->m_Comm.Rank() > 0)
     {
-        a->m_Comm.Recv(
-            &m_DataPos, 1, a->m_Comm.Rank() - 1, 0,
-            "Chain token in BP5Writer::WriteData_EveryoneWrites_Async");
+        a->m_Comm.Recv(&m_DataPos, 1, a->m_Comm.Rank() - 1, 0,
+                       "Chain token in BP5Writer::WriteData_EveryoneWrites_Async");
     }
 
     // align to PAGE_SIZE
-    m_DataPos +=
-        helper::PaddingToAlignOffset(m_DataPos, m_Parameters.StripeSize);
+    m_DataPos += helper::PaddingToAlignOffset(m_DataPos, m_Parameters.StripeSize);
     m_StartDataPos = m_DataPos;
 
     if (a->m_Comm.Rank() < a->m_Comm.Size() - 1)
     {
         uint64_t nextWriterPos = m_DataPos + Data->Size();
-        a->m_Comm.Isend(
-            &nextWriterPos, 1, a->m_Comm.Rank() + 1, 0,
-            "Chain token in BP5Writer::WriteData_EveryoneWrites_Async");
+        a->m_Comm.Isend(&nextWriterPos, 1, a->m_Comm.Rank() + 1, 0,
+                        "Chain token in BP5Writer::WriteData_EveryoneWrites_Async");
     }
 
     m_DataPos += Data->Size();
@@ -266,8 +252,7 @@ void BP5Writer::WriteData_EveryoneWrites_Async(format::BufferV *Data,
         m_AsyncWriteInfo->comm_chain = a->m_Comm.GroupByShm();
         m_AsyncWriteInfo->rank_chain = m_AsyncWriteInfo->comm_chain.Rank();
         m_AsyncWriteInfo->nproc_chain = m_AsyncWriteInfo->comm_chain.Size();
-        m_AsyncWriteInfo->tokenChain =
-            new shm::TokenChain<uint64_t>(&m_AsyncWriteInfo->comm_chain);
+        m_AsyncWriteInfo->tokenChain = new shm::TokenChain<uint64_t>(&m_AsyncWriteInfo->comm_chain);
     }
     else
     {
@@ -285,8 +270,7 @@ void BP5Writer::WriteData_EveryoneWrites_Async(format::BufferV *Data,
     m_AsyncWriteInfo->flagRush = &m_flagRush;
     m_AsyncWriteInfo->lock = &m_AsyncWriteLock;
 
-    if (m_ComputationBlocksLength > 0.0 &&
-        m_Parameters.AsyncWrite == (int)AsyncWrite::Guided)
+    if (m_ComputationBlocksLength > 0.0 && m_Parameters.AsyncWrite == (int)AsyncWrite::Guided)
     {
         m_AsyncWriteInfo->inComputationBlock = &m_InComputationBlock;
         m_AsyncWriteInfo->computationBlocksLength = m_ComputationBlocksLength;
@@ -294,10 +278,8 @@ void BP5Writer::WriteData_EveryoneWrites_Async(format::BufferV *Data,
         {
             m_AsyncWriteInfo->deadline = m_ComputationBlocksLength;
         }
-        m_AsyncWriteInfo->expectedComputationBlocks =
-            m_ComputationBlockTimes; // copy!
-        m_AsyncWriteInfo->currentComputationBlocks =
-            &m_ComputationBlockTimes; // ptr!
+        m_AsyncWriteInfo->expectedComputationBlocks = m_ComputationBlockTimes; // copy!
+        m_AsyncWriteInfo->currentComputationBlocks = &m_ComputationBlockTimes; // ptr!
         m_AsyncWriteInfo->currentComputationBlockID = &m_ComputationBlockID;
 
         /* Clear current block tracker now so that async thread does not get
@@ -318,8 +300,8 @@ void BP5Writer::WriteData_EveryoneWrites_Async(format::BufferV *Data,
         m_AsyncWriteInfo->currentComputationBlockID = nullptr;
     }
 
-    m_WriteFuture = std::async(
-        std::launch::async, AsyncWriteThread_EveryoneWrites, m_AsyncWriteInfo);
+    m_WriteFuture =
+        std::async(std::launch::async, AsyncWriteThread_EveryoneWrites, m_AsyncWriteInfo);
 
     // At this point modifying Data in main thread is prohibited !!!
 
@@ -335,9 +317,8 @@ void BP5Writer::WriteData_EveryoneWrites_Async(format::BufferV *Data,
         }
         if (a->m_Comm.Rank() == 0)
         {
-            a->m_Comm.Recv(
-                &m_DataPos, 1, a->m_Comm.Size() - 1, 0,
-                "Chain token in BP5Writer::WriteData_EveryoneWrites_Async");
+            a->m_Comm.Recv(&m_DataPos, 1, a->m_Comm.Size() - 1, 0,
+                           "Chain token in BP5Writer::WriteData_EveryoneWrites_Async");
         }
     }
 }
